Video 5
Used Tiled to create a map, then looped through every square in the map looking for items from the "blue" layer. When this map is loaded, the blue boxes are surrounded by box2d boxes, which allow the player to interact with them.

The bits are set so that the player can currently only jump on blue. To jump on the other boxes (not yet set), the player will have to change the bitMask to go to RED or GREEN.